Removed wedge for redesign
Wedge to be slotted between the weir and the open barrel is being changed to a flat piece of tin or sheeting which will lay on the open top lip of the barrel and be anchored with two concrete bolts into weir face. Uploading 20 mesh stainless produces error. File too large. 100 MB. Mesh is too complex. Will try to create image of mesh but not a complete file requiring so much memory.